+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+### Changed
+- The behavior of whether a JavaScript number is converted into Python `float` or `int` has changed.
+    - Before: Numbers that are valid 32-bit signed/unsigned integers (`i32`/`u32`) are converted into `int`. Others are converted into `float`.
+    - After: Non-whole numbers are converted into `float`. Whole-number conversion is controlled by the parameter `integer_conversion`. The valid values are, in the order of increasing aggressiveness:
+        - `never`: All numbers are converted into `float`.
+        - `i32`: Only valid 32-bit integers are converted into `int`. This seems to be the common behavior of other embedded JavaScript engines.
+        - `safe`: Whole numbers within the [safe-integer range](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number/isSafeInteger) are converted into `int`. This is the default.
+        - `aggressive`: Even [unsafe integers](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number/MAX_SAFE_INTEGER) are converted into `int`.
 
 ## [0.4.0] - 2024-03-03
 ### Added